Move code to graph dir.
