+++ cpp/dolfinx/io/VTKWriter.cpp
@@ -14,7 +14,7 @@
 #include <dolfinx/function/FunctionSpace.h>
 #include <dolfinx/la/PETScVector.h>
 #include <dolfinx/la/utils.h>
-#include <dolfinx/mesh/Connectivity.h>
+#include <dolfinx/graph/AdjacencyGraph.h>
 #include <dolfinx/mesh/CoordinateDofs.h>
 #include <dolfinx/mesh/Mesh.h>
 #include <dolfinx/mesh/MeshEntity.h>
