--- cpp/dolfinx/mesh/Connectivity.cpp
@@ -1,13 +1,0 @@
-// Copyright (C) 2006-2019 Anders Logg and Garth N. Wells
-//
-// This file is part of DOLFINX (https://www.fenicsproject.org)
-//
-// SPDX-License-Identifier:    LGPL-3.0-or-later
-
-#include "Connectivity.h"
-#include <boost/functional/hash.hpp>
-#include <sstream>
-
-using namespace dolfinx;
-using namespace dolfinx::mesh;
-
