+++ cpp/dolfinx/mesh/MeshFunction.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "Connectivity.h"
+#include <dolfinx/graph/AdjacencyGraph.h>
 #include "Geometry.h"
 #include "Mesh.h"
 #include "MeshEntity.h"
